Create README.md
TASK: Draw a shaded circle in the center of the canvas.
Organize an arrow key handler. If the center of the circle "comes out" beyond the boundaries of the canvas, then the circle should appear from the opposite side. At the bottom of the screen, place a label with the coordinates of the center of the circle in the user coordinate system. Also make sure that when
you click on the 1,2,3 buttons, the circle changes
the fill color to red, blue and green, respectively. And when
you press the c button, the circle changed color to an arbitrary one, and when you press the p button, the circle
moved to an arbitrary position on
the screen.